User upload picture function added

+++ app/Http/Controllers/PostController.php
@@ -8,7 +8,9 @@ use App\Http\Requests\PostRequest;
 use App\Location;
 use App\Photo;
 use App\Post;
+use App\User;
 use Carbon\Carbon;
+use Gate;
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\Session;
 use Illuminate\Support\Facades\Storage;
@@ -18,6 +20,8 @@ class PostController extends Controller
 {
     public function createPost()
     {
+        $this->authorize('create', Post::class);
+
         $categories = Category::all()->pluck('name', 'id');
         $countries = Country::all()->pluck('name', 'id');
         return view('posts.create', compact('countries', 'categories'));
@@ -25,6 +29,8 @@ class PostController extends Controller
 
     public function publishPost(PostRequest $request)
     {
+        $this->authorize('create', Post::class);
+
         $attraction = strtolower($request->attraction);
         $address = strtolower($request->address);
         $city = strtolower($request->city);
@@ -43,14 +49,14 @@ class PostController extends Controller
 
         if ($post->save()) {
             $img = $request->file('photo');
-            $name = $post->id . 'at' . time() . $img->getClientOriginalName();
+            $name = time() . $img->getClientOriginalExtension();
             $imgPath = Storage::putFileAs('public/img/posts', $img, $name);
             $photo = new Photo([
                 'post_id' => $post->id,
                 'imgPath' => $imgPath
             ]);
             if ($photo->save()) {
-                return redirect()->route('post.view', ['id' => $post->id]);
+                return redirect()->route('post.view', $post);
             } else {
                 return redirect()->route('home');
             }
@@ -61,12 +67,16 @@ class PostController extends Controller
 
     public function viewPost(Post $post)
     {
-        $user = Session::has('user') ? Session::get('user') : null;
-        return view('posts.view', compact('post', 'user'));
+        $comments = $post->comments;
+        $user = User::find($post->user_id);
+        return view('posts.view', compact('post', 'comments', 'user'));
     }
 
     public function editPost(Post $post)
     {
+        $this->authorize('update', $post);
+//        abort_unless(Gate::allows('update', $post), 403);
+
         $categories = Category::all()->pluck('name', 'id');
         $countries = Country::all()->pluck('name', 'id');
         $category = Category::find($post->category_id);
@@ -77,6 +87,8 @@ class PostController extends Controller
 
     public function updatePost(PostRequest $request, Post $post)
     {
+        abort_unless(Gate::allows('update', $post), 403);
+
         $attraction = strtolower($request->attraction);
         $address = strtolower($request->address);
         $city = strtolower($request->city);
@@ -91,36 +103,12 @@ class PostController extends Controller
         ]);
 
         return redirect()->route('post.view', ['id' => $post->id]);
-//        $user = Session::get('user');
-//        $post = new Post([
-//            'user_id' => $user['id'],
-//            'title' => strtolower($request->title),
-//            'content' => $request->postContent,
-//            'rate' => null,
-//            'location_id' => $locationId,
-//            'category_id' => $request->category
-//        ]);
-//
-//        if ($post->save()) {
-//            $img = $request->file('photo');
-//            $name = $post->id . 'at' . time() . $img->getClientOriginalName();
-//            $imgPath = Storage::putFileAs('public/img/posts', $img, $name);
-//            $photo = new Photo([
-//                'post_id' => $post->id,
-//                'imgPath' => $imgPath
-//            ]);
-//            if ($photo->save()) {
-//                return redirect()->route('post.view', ['id' => $post->id]);
-//            } else {
-//                return redirect()->route('home');
-//            }
-//        } else {
-//            return redirect()->route('home');
-//        }
     }
 
     public function removePost(Post $post)
     {
+        $this->authorize('delete', $post);
+
         $posts = Post::find($post->location_id);
         if (count($posts) == 1) {
             Location::destroy($post->location_id);
